# server/src/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the scrumlr backend server.",
        "title": "Scrumlr backend",
        "termsOfService": "https://scrumlr.io/terms",
        "contact": {
            "email": "info@scrumlr.io"
        },
        "license": {
            "name": "MIT",
            "url": "https://github.com/inovex/scrumlr.io/blob/main/LICENSE"
        },
        "version": "5.4.0"
    },
    "paths": {
...
        "/templates": {
            "get": {
                "description": "Get all board templates for a user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "board templates"
                ],
                "summary": "Get all board templates",
                "parameters": [
                    {
                        "type": "string",
                        "description": "jwt token to authenticate",
                        "name": "Cookie",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/boardtemplates.BoardTemplateFull"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/common.APIError"
                        }
                    },
                    "429": {
                        "description": "Too Many Requests"
                    }
                }
            },
            "post": {
                "description": "Create a board template",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "board templates"
                ],
                "summary": "Create a board template",
                "parameters": [
                    {
                        "type": "string",
                        "description": "jwt token to authenticate",
                        "name": "Cookie",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Board template to create",
                        "name": "boardtemplate",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/boardtemplates.CreateBoardTemplateRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/boardtemplates.BoardTemplate"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/common.APIError"
                        }
                    },
                    "429": {
                        "description": "Too Many Requests"
                    }
                }
            }
        },
...
        "/templates/{id}": {
            "get": {
                "description": "Get a board template by its id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "board templates"
                ],
                "summary": "Get a board template",
                "parameters": [
                    {
                        "type": "string",
                        "description": "jwt token to authenticate",
                        "name": "Cookie",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Id of the template to get",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/boardtemplates.BoardTemplate"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/common.APIError"
                        }
                    },
                    "429": {
                        "description": "Too Many Requests"
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/common.APIError"
                        }
                    }
                }
            },
            "put": {
                "description": "Update a board template for a user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "board templates"
                ],
                "summary": "Update a board template",
                "parameters": [
                    {
                        "type": "string",
                        "description": "jwt token to authenticate",
                        "name": "Cookie",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Id of the template to update",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Board template to update",
                        "name": "boardtemplate",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/boardtemplates.BoardTemplateUpdateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/boardtemplates.BoardTemplate"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/common.APIError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/common.APIError"
                        }
                    },
                    "429": {
                        "description": "Too Many Requests"
                    }
                }
            },
            "delete": {
                "description": "Delete a board template for a user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "board templates"
                ],
                "summary": "Delete a board template",
                "parameters": [
                    {
                        "type": "string",
                        "description": "jwt token to authenticate",
                        "name": "Cookie",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Id of the template to update",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/common.APIError"
                        }
                    },
                    "429": {
                        "description": "Too Many Requests"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "boardtemplates.BoardTemplate": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "creator": {
                    "description": "The board template creator id",
                    "type": "string"
                },
                "description": {
                    "description": "Description of the board template",
                    "type": "string"
                },
                "favourite": {
                    "description": "The favourite status of the template",
                    "type": "boolean"
                },
                "id": {
                    "description": "The board template id",
                    "type": "string"
                },
                "modifiedAt": {
                    "type": "string"
                },
                "name": {
                    "description": "The board template name",
                    "type": "string"
                }
            }
        },
        "boardtemplates.BoardTemplateFull": {
            "type": "object",
            "properties": {
                "columns": {
                    "description": "Board templates associated column templates",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/columntemplates.ColumnTemplate"
                    }
                },
                "template": {
                    "$ref": "#/definitions/boardtemplates.BoardTemplate"
                }
            }
        },
        "boardtemplates.BoardTemplateUpdateRequest": {
            "type": "object",
            "properties": {
                "description": {
                    "description": "Description of the board template",
                    "type": "string"
                },
                "favourite": {
                    "description": "The favourite status of the template",
                    "type": "boolean"
                },
                "id": {
                    "description": "The board template id",
                    "type": "string"
                },
                "name": {
                    "description": "The board template name",
                    "type": "string"
                }
            }
        },
        "boardtemplates.CreateBoardTemplateRequest": {
            "type": "object",
            "properties": {
                "columnTemplates": {
                    "description": "The column templates to create for the board template.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/columntemplates.ColumnTemplateRequest"
                    }
                },
                "creator": {
                    "description": "The user who creates the template",
                    "type": "string"
                },
                "description": {
                    "description": "Description of the board template",
                    "type": "string"
                },
                "favourite": {
                    "description": "The favourite status of the template",
                    "type": "boolean"
                },
                "name": {
                    "description": "The name of the board template.",
                    "type": "string"
                }
            }
        },
...
        "columntemplates.ColumnTemplate": {
            "type": "object",
            "properties": {
                "color": {
                    "description": "The column template color.",
                    "allOf": [
                        {
                            "$ref": "#/definitions/common.Color"
                        }
                    ]
                },
                "description": {
                    "description": "The description of a board template column.",
                    "type": "string"
                },
                "id": {
                    "description": "The column template id.",
                    "type": "string"
                },
                "index": {
                    "description": "The column template rank.",
                    "type": "integer"
                },
                "name": {
                    "description": "The column template name.",
                    "type": "string"
                },
                "template": {
                    "description": "The board template id, that this column template is bound to.",
                    "type": "string"
                },
                "visible": {
                    "description": "The column template visibility.",
                    "type": "boolean"
                }
            }
        },
        "columntemplates.ColumnTemplateRequest": {
            "type": "object",
            "properties": {
                "color": {
                    "description": "The column template color to set.",
                    "allOf": [
                        {
                            "$ref": "#/definitions/common.Color"
                        }
                    ]
                },
                "description": {
                    "description": "The columnTemplate description to set.",
                    "type": "string"
                },
                "index": {
                    "description": "Sets the index of this column template in the sort order.",
                    "type": "integer"
                },
                "name": {
                    "description": "The column template name to set.",
                    "type": "string"
                },
                "visible": {
                    "description": "Sets whether this column template should be visible to regular participants.\nThe default value on creation is 'false'.",
                    "type": "boolean"
                }
            }
        },
        "common.APIError": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
        "common.Color": {
            "type": "string",
            "enum": [
                "backlog-blue",
                "goal-green",
                "value-violet",
                "online-orange",
                "planning-pink",
                "poker-purple",
                "yielding-yellow"
            ],
            "x-enum-varnames": [
                "ColorBacklogBlue",
                "ColorGoalGreen",
                "ColorValueViolet",
                "ColorOnlineOrange",
                "ColorPlanningPink",
                "ColorPokerPurple",
                "ColorYieldingYellow"
            ]
        },
...
    }
}